gameui mapScreenDrawHud 97.50->97.60: cast drawTexture alpha arg to (u8)panelAlpha
drawTexture's param is int; passing raw s16 panelAlpha emitted extsh+mr and forced MWCC
to keep both s16 and a CSE'd u8 live. Matching the adjacent drawScaledTexture calls'
(u8) cast emits clrlwi directly, dropping 2 instrs.

+++ src/main/dll/dll_0000_gameui.c
@@ -7344,7 +7344,7 @@ void mapScreenDrawHud(int p1, int p2, int p3)
         panelY = *(s16*)(gTextBoxes + 0x196);
         height = revealedHeight;
         width = (s16) * (u16*)(gTextBoxes + 0x182);
-        drawTexture(((HudTextures*)hudTextures)->tex28, (f32)(panelX - 5), (f32)(panelY - 5), panelAlpha,
+        drawTexture(((HudTextures*)hudTextures)->tex28, (f32)(panelX - 5), (f32)(panelY - 5), (u8)panelAlpha,
                     0x100);
         drawScaledTexture(((HudTextures*)hudTextures)->tex34, panelX, (f32)(panelY - 5), (u8)panelAlpha, 0x100,
                           width, 5, 0);
